Rearrange processing at end to avoid reprocessing
The code to insert a "dummy" event at the end was intended
to mark the data file with the time of the last checked dwell.
With the previous organization, it was only updating the file
if there were new high-background events, leading to reprocessing
all the dwells since the last high-background event.  This
was not the intent, so a small rearrange/reorg fixes that issue.